Revert "phase - 8 -- current stock count and pricing datatype updated"
This reverts commit 25c76910ba16ae853539935d4b4db12f0b9e94bb.

+++ app/lib/data/models/product.dart
@@ -12,10 +12,10 @@ abstract class Product with _$Product {
     required String brand,
     required String categoryId,
     required int minimumStock,
-    required int stock, // Current stock quantity count
-    required double costPrice, // Cost price in Rupees (float)
-    required double sellingPrice, // Selling price in Rupees (float)
-    required double mrp, // Maximum Retail Price in Rupees (float)
+    required int stock, // Current stock quantity
+    required int costPrice, // Cost price in paise (₹ × 100)
+    required int sellingPrice, // Selling price in paise (₹ × 100)
+    required int mrp, // Maximum Retail Price in paise (₹ × 100)
     String? imageUrl,
     String? description,
   }) = _Product;
